v5.5.1 merge, part 2

v5.5.1 merge part 3

v5.5.1 merge part 4

v5.5.1 merge part 5

v5.5.1 merge part 6

v5.5.1 merge part 7

v5.5.1 merge part 8

v5.5.1 merge part 9

v5.5.1 merge part 10

+++ genlib/setline.c
@@ -10,7 +10,7 @@
 
 
 int
-setline(float *p, short n_args, int length, double *array)
+setline(double *p, int n_args, int length, double *array)
 {
    double increm;
    int i, j, k, i2, points;
